pyproject.toml

[project]

# 需要更改发布到pypi的包名时改这里
name = "ipa-python-kit"
dynamic = ["version"]
description = "kit for python"
authors = [{ name = "xdewx", email = "present150608@sina.com" }]
keywords = ["python", "kit", "lib", "util"]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "deprecated>=1.3.1",
    "pint>=0.21.1",
    "pydantic>=2.10.6",
    "pyhumps>=3.8.0",
    "typer>=0.20.1",
]

[project.optional-dependencies]
system = ["psutil>=7.2.1"]
i18n = ["python-i18n>=0.3.9"]
log = ["colorlog>=6.10.1"]
hmr = ["python-hmr>=0.3.0", "watchfiles>=0.24.0"]
date = ["python-dateutil>=2.9.0.post0"]
websocket = ["websockets>=13.1"]
visualization = ["matplotlib>=3.7.5", "pyecharts>=2.0.9"]
orm = ["alembic>=1.14.1", "sqlmodel>=0.0.29"]
network = ["networkx>=3.1"]

all = [
    "ipa-python-kit[system]",
    "ipa-python-kit[i18n]",
    "ipa-python-kit[log]",
    "ipa-python-kit[hmr]",
    "ipa-python-kit[date]",
    "ipa-python-kit[websocket]",
    "ipa-python-kit[visualization]",

]
des = ["simpy>=4.1.1"]

# 使用uv_build时
# [build-system]
# requires = ["uv_build>=0.7.4,<0.8.0"]
# build-backend = "uv_build"

# [tool.uv.build-backend]
# module-root = "src"
# module-name = "ipa"

[build-system]
requires = ["setuptools>=80.0", "setuptools-scm>=9"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_file = "src/ipa/version.py"
version_scheme = "guess-next-dev"
# version_scheme = "only-version"
# local_scheme = "node-and-date"
local_scheme = "no-local-version"

[tool.setuptools.package-dir]
"" = "src"

# # SDK实现方式1：路径映射，便于用户通过有辨识度的方式导入模块而不是使用sdk或重命名，本地开发时无法实现代码提示
# my_sdk = "src/my_sdk"

# [tool.setuptools]
# packages = ["ipa"]

# SDK实现方式2：通过自动发现打包模块，结合sdk重命名，可以支持代码提示
[tool.setuptools.packages]
find = { where = ["src"], include = ["ipa"] }

[tool.uv]
index = [
    { "name" = "tencent", "url" = "https://mirrors.cloud.tencent.com/pypi/simple/" },
    { "name" = "pypi", "url" = "https://pypi.org/simple/" },
]

[dependency-groups]
dev = [
    "build>=1.2.2",
    "twine>=4.0.2",
    "black==24.8.0",
    "commitlint==1.3.0",
    "darker==2.1.1",
    "pre-commit==3.5.0",
    "pytest==8.3.5",
    "ruff==0.13.1",
    "setuptools-scm>=9.2.2",
    "pyinstaller>=6.18.0",
    "pytest-cov>=7.1.0",
]
